Destroy properly droplets that are still "new"
The Digital Ocean API does not destroy droplets that are not
yet active. In cases when tests are run with concurrency and
with the option of always destroying instances, when a test
fails, droplets that are still "new" are not properly destroyed
and keep on running.

This fixes the issue by checking the status of droplets upon
destruction: if the droplet is still "new", the destruction
is requested again as long as the droplet is still running.